update audio service

+++ mobile/lib/src/audio_service.dart
@@ -1,7 +1,9 @@
 import 'dart:async';
 import 'dart:io';
 import 'dart:typed_data';
+import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter_sound/flutter_sound.dart';
+import 'package:http/http.dart' as http;
 import 'package:path_provider/path_provider.dart';
 import 'package:permission_handler/permission_handler.dart';
 
@@ -10,14 +12,13 @@ class AudioService {
   bool _isInitialized = false;
   String? _tempFilePath;
 
-  AudioService() {
-    _init();
-  }
-
   Future<void> _init() async {
-    final status = await Permission.microphone.request();
-    if (status != PermissionStatus.granted) {
-      throw Exception('Microphone permission not granted');
+    if (!kIsWeb) {
+      final status = await Permission.microphone.request();
+      print(status);
+      if (status != PermissionStatus.granted) {
+        throw Exception('Microphone permission not granted');
+      }
     }
     await _recorder.openRecorder();
     _isInitialized = true;
@@ -28,28 +29,44 @@ class AudioService {
       await _init();
     }
 
-    final tempDir = await getTemporaryDirectory();
-    _tempFilePath = '${tempDir.path}/flutter_sound_example.wav';
+    if (kIsWeb) {
+      await _recorder.startRecorder(
+        codec: Codec.opusWebM,
+      );
 
-    await _recorder.startRecorder(
-      toFile: _tempFilePath,
-      codec: Codec.pcm16WAV,
-    );
+      await Future.delayed(Duration(milliseconds: durationMs ?? 3000));
 
-    await Future.delayed(Duration(milliseconds: durationMs ?? 3000));
+      final url = await _recorder.stopRecorder();
 
-    await _recorder.stopRecorder();
-
-    if (_tempFilePath != null) {
-      final file = File(_tempFilePath!);
-      if (await file.exists()) {
-        final bytes = await file.readAsBytes();
-        await file.delete();
-        _tempFilePath = null;
-        return bytes;
+      if (url != null) {
+        final response = await http.get(Uri.parse(url));
+        return response.bodyBytes;
       }
+      return Uint8List(0);
+    } else {
+      final tempDir = await getTemporaryDirectory();
+      _tempFilePath = '${tempDir.path}/flutter_sound_example.wav';
+
+      await _recorder.startRecorder(
+        toFile: _tempFilePath,
+        codec: Codec.pcm16WAV,
+      );
+
+      await Future.delayed(Duration(milliseconds: durationMs ?? 3000));
+
+      await _recorder.stopRecorder();
+
+      if (_tempFilePath != null) {
+        final file = File(_tempFilePath!);
+        if (await file.exists()) {
+          final bytes = await file.readAsBytes();
+          await file.delete();
+          _tempFilePath = null;
+          return bytes;
+        }
+      }
+      return Uint8List(0);
     }
-    return Uint8List(0);
   }
 
   void dispose() {
